InputForm, CreateTr 컴포넌트 분리
