+++ src/App.css
@@ -100,11 +100,11 @@ table td {
 }
 
 tbody > tr:nth-child(2n + 1) {
-  background-color: #d6e6f5;
+  background-color: #ebf3fa;
 }
 
 tbody > tr:nth-child(2n) {
-  background-color: #d1e6d1;
+  background-color: #e8f3e8;
 }
 
 td > button {
